+++ definitions/docker_build_nobase.rb
@@ -1,16 +1,19 @@
 define :docker_build_nobase do
 
-  #class Chef::Resource
-  #  include DockerWrapper
-  #end
-
+  class Chef::ResourceDefinitionList
+    include DockerHelper
+  end
+  
   enable = params[:enable_service]
+  base_image_exists = DockerWrapper::Image.exists?("#{params[:initial_base_image_name]}:#{params[:initial_base_image_tag]}")
+  image_exists = DockerWrapper::Image.exists?("#{params[:project_image_name]}:#{params[:project_image_tag]}")
 
   ## get the initial base image
   docker_deploy_image "#{params[:initial_base_image_name]}_pull" do
     name params[:initial_base_image_name]
     tag params[:initial_base_image_tag]
     action enable ? :pull_if_missing : :remove_if_unused
+    not_if { enable and base_image_exists }
   end
   
   ## get project specific image (if available)
@@ -19,6 +22,7 @@ define :docker_build_nobase do
     tag params[:project_image_tag]
     action enable ? :pull_if_missing : :remove_if_unused
     ignore_failure true
+    not_if { enable and image_exists }
   end
 
   ## otherwise build service specific image
@@ -35,7 +39,7 @@ define :docker_build_nobase do
     chef_admin_key params[:chef_admin_key]
     docker_build_commands params[:docker_build_commands]
     action :build_if_missing
-    #only_if { enable and DockerWrapper::Image.get("#{params[:initial_base_image_name]}:#{params[:initial_base_image_tag]}").exists? }
+    only_if { enable and base_image_exists }
   end
 
   ## push
