+++ definitions/docker_push.rb
@@ -1,10 +1,12 @@
 define :docker_push do
 
-  class Chef::Resource
-    include DockerWrapper
+  class Chef::ResourceDefinitionList
+    include DockerHelper
   end
 
   enable = params[:enable_service]
+  base_image_exists = DockerWrapper::Image.exists?("#{params[:initial_base_image_name]}:#{params[:initial_base_image_tag]}")
+  image_exists = DockerWrapper::Image.exists?("#{params[:project_image_name]}:#{params[:project_image_tag]}")
 
   ## auth
 
@@ -13,7 +15,7 @@ define :docker_push do
     name params[:project_base_image_name]
     tag params[:project_base_image_tag]
     action :push
-    only_if { enable and get_exists?("#{params[:project_base_image_name]}:#{params[:project_base_image_tag]}") }
+    only_if { enable and base_image_exists }
   end
   
   ## revision image
@@ -21,6 +23,6 @@ define :docker_push do
     name params[:project_image_name]
     tag params[:project_image_tag]
     action :push
-    only_if { enable and get_exists?("#{params[:project_image_name]}:#{params[:project_image_tag]}") }
+    only_if { enable and image_exists }
   end
 end
